1018 1182 1189 1261 1303 1520 1600 1620 1920 1926 1938 1941 2178 2206 2234 2423 2573 2589 2617 2636 2638 2718 2753 2776 2966 2991 3055 3184 3187 3197 4179 5214 5427 6087 7576 9328 10815 10828 10845 10866 10988 11403 11723 12904 13023 13246 13430 13459 13565 13908 14492 14502 14716 14940 15917 15963 16139 16171 16172 16199 16916 16933 16940 16946 16954 16964 17086 17210 17836 20056 21591 21771 23899 23900 23970 24053 24062 24092 24313 24314 24315 24368 24369 24370 24400 25238 25240 25501 25512 25629 28278 수정

+++ rust/src/02/2966.rs
@@ -9,14 +9,14 @@ fn main() {
         let score: i32 = input
             .char_indices()
             .map(|(i, ch)| {
-                i32::from(match id {
+                (match id {
                     "Adrian" => matches!((i % 3, ch), (0, 'A') | (1, 'B') | (2, 'C')),
                     "Bruno" => matches!((i % 4, ch), (0, 'B') | (1, 'A') | (2, 'B') | (3, 'C')),
                     "Goran" => {
                         matches!((i % 6, ch), (0 | 1, 'C') | (2 | 3, 'A') | (4 | 5, 'B'))
                     }
                     _ => unreachable!(),
-                })
+                }) as i32
             })
             .sum();
         max_score = score.max(max_score);
